Fix gh-813 Add stylesheet on Thor Status
The existing code retrieves the tree branch from Status/servers/
for a thor cluster and convert the branch into XML for display.
This fix adds a ws_topology method and a stylesheet for viewing
thor status based on the tree branch. Major data items of the
tree branch are read and displayed inside a html form. From the
form, a user may click a link to view the XML; a user may click
a link to open thor log file; and, if a WU is running on that
thor process, a user may click a link to view WU Details page.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>