Fix: searching logical files: hthor cluster not available
The problem in issue ln-129 is that ECLwatch should not show
hthor__myeclagent as the name for hthor. The hthor__myeclagent
is the name of the group associated with 'hthor' target
cluster. The group name is used for Spray and it should not be
presented to the user in ECLwatch. Based on Jake's suggestion,
this fix replaces the group names using target cluster names.
The changes include: (1) display target cluster names (such as
thor, roxie, hthor) in three Spray forms; (2) after a target
cluster name is selected, the queue name associated with the
target cluster is used to call a spray function; (3) display
target cluster names inside the Searching Logical File form;
(4) after a target cluster name is selected, the group name
associated with the target cluster is used to Searching
Logical File; (5) for each logical file inside the results of
Searching (and Browse) Logical File, replace the group name
with the target cluster name; (6) In the Logical File Details
page, replace the group name with the target cluster name.
(7) for searching (and displaying) DFU workunit(s), existing
code does switch from target cluster name to the group name
when retrieving and switch it back for display, if the cluster
is a thor cluster or a roxie cluster. This fix implements the
same logical for hthor cluster.

In addition, the new code uses the workunit lib methods (such
as getEnvironmentThorClusterNames(), etc) to retrieve the
names of target clusters. A new workunit lib method is added:
getEnvironmentRoxieClusterNames().

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>