Fix gh-709 StandAlone data file location fix
If running stand alone with a dali connection (config file not available),
the data directory defaults to /c$/thordata. This is incorrect as it
should be querying that location from dali.
Fix the problem by not setting default base to "hthor", therefore enabling
the conditional code in defdesc.cpp "loadDefaultBases" to query from dali.
Also, fixed a bug in dautils where allowospath(bool) was ignoring the flag
and always setting allow to true.

Signed-off-by: William Whitehead <william.whitehead@lexisnexis.com>