gh-439 hpcc user should be a system account
The hpcc user should be a system account, and should have a locked password.
useradd -r is part of the LSB spec and should be ok on all distros we support
(was added to debian in 2006 and RH before that).

Without this fix desktop users can log in to the hpcc account without any
authentication.

Fixes gh-439.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>