Fix gh-914, Ensure CSV headers skipped
It was assumed the csv header lines, existed on the 1st node, but they might
exist on any, and/or span nodes.
If a csv input has a header, it will now mean that the nodes must wait until
signalled by the previous node before proceeding. Once the header lines are
consumed, the consumer will signal all remaining nodes immediately.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>