Protect against child query in master context
Child queries should not be generated such that they are fired from helper
methods called in Thor Master. The only known way this can happen is via
NOHOIST. That resulted in an obscure crash, this change detects and gives a
sensible error and context.
Regression test loop2.ecl was an example that provoked this condition, which
has now been marked //nothorlcr

Also:
1) improved abort handling in sort
2) improved exception messages in some contexts
3) Fix a possible exception leak

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>