Transaction refactoring - create super file
Adding super-files within transaction. Locking and unlocking within
the transaction, making sure the file exists and onle unlock on
rollback if it wasn't. Transaction is now compulsory, even if
not active, which will trigger an auto-commit. A local transaction
is created in case the pointer is NULL, but eventually, it'll
assertex on transaction existing.

This commit narrowes the windows to which a super file stays in the
file system (from infinite to just between transaction's lock and
unlock), but the gap is still there. There is a FIXME to that effect.

Also, the action's lock/unlock add and remove the entry from the
filesystem, which in case of retry is inefficient, but we must
know if the addition will fail before we lock the other actions.
Would be good to have a solution to that, as well.