Allow max cores (for sort) to be configured
gh-1099 - Add thor configuration option 'maxActivityCores' to allow the max number of
cpu cores to be specified per activity. Currently only used by sorting
activities.  The default is all available cores.
The option is available as a HINT in ECL, as a #option and as
a thor configuration option. With precedence taken in that order.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>