Fix compile errors on Tilera/Mac versions of gcc
Default gcc installed with XCode on OSX Lion is
i686-apple-darwin11-llvm-gcc-4.2 (GCC) 4.2.1
(Based on Apple Inc. build 5658) (LLVM build 2336.1.00)

This seems to have some issues with virtual functions that have ... params
if overridden, if base class is not the first. Easiest way to avoid these
issues (which also affected Tilera build) is to swap the order of the base
classes in a couple of cases.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>