Use GCC atomic operations in more cases
The atomic operations provided by gcc are not specific to linux, and should
be used wherever they are available rather than falling back to the (very
slow) alternative implementations in jiface.cpp.

The alternative implementations in jiface.cpp are also faulty, with the
decrement and increment reversed. This commit fixes that typo.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>