Unpredicable initialization order may core eclcc
The addition of atoms to atom table is split between the initialization
code of hql and hqlcpp libraries, but only hql's initialization code creates
the atom table. Therefore it is important to ensure that it is initialized
first. We appear to have generally got lucky on this in the past, since both
use the same priority, but on Macintosh builds they happen in the wrong order,
causing eclcc to segfault on startup.

Change the priority of hqlcatom.cpp initialization to ensure that it happens
after the initialization in hqlatoms.cpp.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>