Support disabling more functionality in build
To build eclcc in archive mode only (i.e. all that you need for a client)
a lot of components are not needed. By making it easier to disable these
in the cmake script, it's easier to build a mac client system.

While a full build will fail without some of these components, a partial
one such as make eclcc will succeed without most of them.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>