Library resolution failing for non-lowercase libraries
The code in workunit.cpp lowercases all alias names when adding queries
to querysets, but not when looking up a library in a queryset. This means
that a library cannot be located unless its name is all lower case.

Add code to lowercase the name being looked up when resolving a library.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>