Fix graph prepare threading issue
Sub graphs executed by the graph executor thread could overlap the main
dependency walking thread, resulting in the same graph being prepared on two
different threads. Normally this went unoticed, but could cause problems if
a conditional was being followed and being reevaluated at the same time.
Also fix a related, exposed issues - if an OUTPUT,UPDATE was part of a set
of subgraphs and the last one to execute and already up to date, then the
graph executor would wait indefinetely for it to finish

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>