Very rare race condition in Roxie memory manager
Observed an assert failure running RoxieMem self tests, caused by marking
a block as ok to free just after it has been put back into use.

There's not really a god way to avoid this race, and the benefit of using blocks
that are about to be freed is marginal at best (there's a very small window it can
even happen in) so the simplest fix is to remove that code.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>