gh-1160 Compile errors with glibc 2.14
In glibc 2.14 the definition of __malloc_init_hook has changed, so
we need to change the code in jdebug that uses it to match.

While testing, I also discovered some code that was incorrectly inside
the #else clause and thus prevented the system from compiling when full
memory hooks were in use.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>