Forbid adding superfiles to themselves
If the file lookup is successful (one way or another), we have a super file.
Thus, we can compare its name to the sub-file name and make sure we're not
adding the super-file to itself. The queryFileName returns a null terminated
string, so strcmp does not incurr in buffer overflow.