Fixes gh-806 Optional default values not present in the environment are marked in different color
Optional default values not present in the environment are marked in red. A new context menu
item has been added for these items to write them to environment. Note that the environment
still needs to be saved after this action for it to be saved to the environment file. The tooltip
has been updated to reflect that this value is not present in the environment.


Signed-off-by: Sridhar Meda <sridhar.meda@lexisnexis.com>