Set initial size for allocator to min not max
Change the initialSize in CachedOutputMetaData as used by allocators, to
the minimum record size not the maximum.
For a variable length record, which has a largish max, a low minimum and
which is frequently small in practice - pre-allocating max can cause the
memory
manager to fragment. Much better to allocate min and expand from there.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>