Fix graph executor, needs to block if running
Previous fix for a race condition in the graph executor (gh-1159) avoided
the race, by preparing inside the executor, but meant the prepare could occur
too early, if a dependency was in flight.
If traversing another sink graph and executing dependencies, the graph
executor needs to block and wait if a dependent graph is already running.
Otherwise prepare() may attempt to use something created by the
dependencies, e.g. a global in a conditinal branch of the graph.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>