Fix gh-1277 problems of Event Scheduler and deschedule
This change fixes three problems related to EclWatch Event Scheduler
link and deschedule function. Problem #1: when Scheduler link on the
left side menu is clicked, Event Scheduled WUs do not show up. Two
bugs are found for this problem. In Scheduleread.cpp, if the Event
Name is an empty string, the CRootScheduleReader should be used. In
ws_workunit dll, a valid cluster name should be input to the
getScheduleWUs(). Problem #2: on Browse Workunits page, Deschedule
button is not activated when Scheduled WUs are selected. The problem
is caused by missing code for setting a WU flag. The flag is
used to indicate that the WU can be descheduled. Problem #3: when
Deschedule button on Workunit Details page is clicked, EclWatch shows
an error message: Invalid Action 'deschdule'. The problem is due to a
missing value: ActionEventDeschedule inside WU action table.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>