DiskGroupAggregate doesn't delete temp asap
A disk group aggregate reading a temp file, will not free the reader early
enough. As a consequence, if it's the last reader of the spill file, it will
try to delete it and may fail if the file system has the file locked.
The temp will still be deleted, but not until the end of all subgraph in the
graph.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>