Add DISTRIBUTED flag to DATASET(N, trans(COUNTER))
This is patch is just adding the keyword and persisting
it in the object (via isDistributed() function), but
it's not being used yet. We need to fix the limits of
getRow() to obey numRows() in the engine before coding
the distributed part.

The fix mentioned above is my next step.