Fix gh-1283 EclWatch Config button not display esp.xml
The problem happens on FireFox. The esp.xml is sent to the
browser. The browser does not display the esp.xml correctly
because XML style sheet is not specified for the esp.xml.
This fix creates a new function which allows an XML style sheet
to be specified before the file is sent to the browser. The fix
also removes the code for displaying in plain text format as an
option because the configuration file is an xml file.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>