Fix gh-1343 Add checkbox “No separator” to Spray csv input
When a user sprays a CSV file using EclWatch, the user should
be able to specify that there is no separator in the CSV file.
This fix adds a checkbox in Spray CSV input page to indicate
No Separator. When the checkbox is selected, the Separator
input field will be disabled.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>