Fix gh-1345 WU Result link not work for Chrome
In EclWatch WU Details page, the links inside the Results
section can be used to view or download the results. Those
links do work for IE and FireFox, but not for Chrome. It is
because the links use our javascript function called 'download'.
That function name makes Chrome confused since each <a> link
has an attribute called 'download'. This fix changes the
javascript function name and the problem is fixed.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>