Change lockProperties API
lockProperties locks all properties in the connection, and not just
the Attr section, which is what was being returned. Now, the method
locks and return a boolean indicating if one needs to refresh its
data because the lock was not acquired clean. Using queryProperties
to get the properties after locking, to make it clear what's being
returned.

TODO: rename queryProperties to queryFileAttr (since that's what's
being returned), but this is a much bigger change (so better in a
separate patch), since there are other queryProperties around that
do different things.