Fix Better display idle servers on Activity page
This fix is for issue #1357. In EclWatch Activity page, the first
server table has a row to display column names. The rest of server
tables do not has the row. If the table of an idle server is not
the first table, the table will be empty and Chrome or Firefox
shrink the table into one dot. This fix adds a message inside the
table: 'No active workunit' and forces the table to be the same
width as the first server table. This fix also changes the font
of server name to be bold for a better display.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>