Roxie memory leaks
Roxie leaks a DataMemoryBuffer per ping received when pingInterval is non-zero.
Also if udpRetryEnabled is false, all outboud packets are leaked (!).

This commit partially addresses gh-1358, but does NOT address the issue of ping
responses causing the retry queue to fill up. Until that is also fixed, it is
probably safest to keep pingInterval set at 0.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>