Add EclDirect to EclWatch
This fix adds a Run Ecl menu under Ecl (Workunits) menu section
inside EclWatch. The menu is linked with EclDirect RunEclEx
method. The RunEclEx is an enhanced version of the existing
RunEcl method. After the menu is selected, a Run Ecl input form is
displayed. A user may type in Ecl code, specify a cluster if
needed, and run the Ecl code. The results of the Ecl code execution
will be displayed in XMl format or table format.

The port of EclDirect service is moved from 8008 to 8010. In
EclWatch navigation panel, the 'ECL Workunits' menu section
is renamed as 'ECL'.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>