Fix gh-622 set defaultCopies and ensureReplicate
Fixes the replicate bug on some files, especially when
run from Thor. This commit does not fix the underlying
issue, since that would require a big change in the naming
system (ie. recognizing types of files, ensuring net
address, filename, etc is always correct and reachable),
but that change is planned and will be done (See #654).