gh-1412 - Make slave script independent of init
When a new node is swapped in, that hasn't previously had a role as a thor
slave, i.e. it is not listed in the environment. It will not be initialized
by the init system, IOW, it will have no component directories, nor any
configgen generated initialization files.
Thor slaves depended on some of these, the directories and a few vars.
Change thor scripts, to make them self-reliant.

Also, avoid needing to provide dali to slave, slave can get dali (and other
settings) after it has registered with master.
+ Fix slavesPerNode wasn't generated correctly if unset and scripts simpler
if always set in master setvars

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>