Compound filter group diskread could lose records
A bug in the end of group marker processing, could cause early termination
of diskread, if a group is filtered out entirely.

A similar bug on non-compount grouped disk read, but I think it could not be
hit - it would have required a empty to group to have been read from disk.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>