Remove CTempActiveTransaction
FileServices' lookup was temporarily activating the transaction
just to turn on the file caching. So far, it did not have side
effects, but some of the failed attempts to refactor the transactional
model in the DFS have hit this problem more than once. Adding a
cache option, if that's the only thing that is needed.

CDFile's detach had to be forced to loose the connection, and that
supports the idea that detach has to be an internal implementation
and probably factored out anyway, since it's no more than a glorified
removeEntry.