Better alignment for Name column in Browse Files by Scope
This fix is based on issue #1631: The name column for the
browse files by scope page is centred when viewed with
chrome. This looks strange.

The default for <td> should be 'Left-aligns content'. For
some reason, Chrome does not follow the default. So, the
column has to be coded into 'Left-aligns content'.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>