Fix segfault when activities stop before even starting
Roxie activities stop on exception, and there is a difference between
abort stop and normal stop. One of the things it that an abort
stop can happen even before its related start took place, because
the exception handling stop chain has no idea how many starts took
place, so it just stops them all. One of the stops in this chain
didn't propagate the 'abort' flag, and stopping for real, expecting
a pointer to be valid, segfaulting.