Fix problems with global priority stepped join.
For global stepped joins (using the priority ordering) it was possible
for the seek list to contain any item too early (which triggered an
assert in the new code).  It may possibly have caused problems in old
builds.

There was another problem where rows could (in very rare situations)
get lost if there were large numbers of matches for a single document.

Fix problem with hint attributes - not allowed on stepped, lost when a
project and table combined, and occasionally duplicated in the graph.

Add hint to roxie to allow the seek readahead to be configured.

Signed-off-by: Gavin Halliday <gavin.halliday@lexisnexis.com>