Change memory size defaults + allow configurable
Allow the global pool to be configurable (per job) and allow it to be unbound,
IOW, allow it consume all available OS memory. Set to 0 to be unbound.
Allow 'largeMem' size to be configurable per job, and make it default to 75%
of global mem.
If global mem unset it will default to 75% of 2GB - this will change to
be automatically based on physical RAM and take into account slavesPerNode
in a future commit.

NB: largeMemSize setting was being completely ignored, due to a typo.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>