Preflight should use new pid file name for multiSlaves
The multiSlaves flag is used by the existing Preflight code
to decide whether PID file names for multiSlaves contain
'_slaveNumber' or not. That caused problems when old version
of environment.xml is used because multiSlaves flag is set
inside the xml file. In this fix, the '_slaveNumber' is
always appended to the PID file names for thor slaves even
when the multiSlaves flag is set. The multiSlaves flag will
only be used for removing swapnode options for legacy system.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>