Allow makeAbsolutePath on non-existent paths
Prior to gh-1605, win/lin would return a absolute path for non-existent file
paths.
In Linux, this was despite realpath returning an error
i.e. the result was still expanded to some extent, in most cases obscuring
the error.

The introduction of the throw if !realpath, broke a few components,
where paths that did not [yet] exist were passed.
[The main cases I saw were via openLogFile], but there were others.

This change allows either it to insist on the path being resolvable, or
attemps to resolve as much as possible, i.e. expand links etc. The latter
being the default.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>