gh-1332 - Reduce dafilesrv inactive client logging
A connected inactive client (>12mins), was being checked every minute
to see if was still alive and if it was "Inactice client" messages for all
files from the client were being logged.
This can be excessive in some situations.

Still want to check per minute, but add extra timeout to control frequency
of warning

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>