Improve implicit field projection for count/exists
- Previously the output record from an activity was required to have at least one
  field, which meant that exists/count arbitrarily "used" the first field.  With
  this change that no longer happens - often reducing the number of fields used.

A record can have a _nonEmpty_ attribute associated with it to ensure
it has a non-zero size (to differenitate it from a skipped row).  The
transform clears the byte (to make memory consitant and ensure spilt
files compress properly).

Serialize/Deserialize use the default classes so they will still work.

Signed-off-by: Gavin Halliday <gavin.halliday@lexisnexis.com>