gh-1646 - Roxie groups only updated on new env.
Thor cluster groups are only updated if the environment is detected to be
different from both the old environment and the live group.
This was disabled for roxie cluster/groups, i.e. the loaded environment
always too precedence.

This change treates roxie cluster group updates in the same way as thor
cluster group updates.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>