gh-1865 Roxie can core when query deployed
A race condition between the control:reload query sent by ecl publish,
and the automatic reload triggered by the dali subscription, can cause
Roxie to core if the notify call is in progress when the QuerySet manager
is destroyed by the control:reload

Add a critical section so that a we never unsubscribe when a notify is
in flight.

Also deleted an unused member variable. Fixes gh-1865

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>