Improve cse inside nodes that can't create temps
If you have a cse within a node that can't have a temporary created
for it e.g., str[1..f()] and other[1..f()].  The no_range is common
but doesn't correspond to an expression that can be generated.  With
this fix f() is spotted as a cse.

Has a dramatic effect on a small subset of the archived queries.  A
prerequisite for another change that is comming up.

Signed-off-by: Gavin Halliday <gavin.halliday@lexisnexis.com>