Roxie core in self-join with prefix match optimization
The prefixjoin.ecl test case in regression suite cores if executed
with a choosen on the output datasets, as prefixed selfjoin was
not correctly handling the case where not all data was pulled.

Code was casting a pointer to a more derived type incorrectly, then
attempting to avoid the resulting issues by copying a valid pointer
back in before any virtual calls that migth crash were made. Yuk!

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>