Remove more traces of (now unnecessary) hack in prefix join
Prefix join code used to rely (unsuccessfully) on having reset the input
to point to a valid object before any calls that might crash because it
pointed to an invalid VMT (!)

This is no longer necessary, but missed removing one bit of the code to do
it. This would not cause any ill effect other than confusing future coders.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>