Disk aggregate and index aggregate could corrupt memory
If the result of a disk aggregate or index aggregate was a variable
size row, roxie heap could get overwritten

Removed the unsafe version of the cloneRow function - which is never
 safe to use in new memory paradigms - to ensure that no other issues
remain. A couple of changes in hthor to avoid the unsafe functions
were unlikely to cause corruptions in practice.

Legacy thor does not appear to use these deprecated functions.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>