Show queries suspended on a given cluster
EclWatch lists queries in a queryset, and shows which are
globally suspended, but we should provide a way of seeing
the state of those queries on a specific cluster. This fix
displays a drop down list of clusters and a drop down list
of query status (Suspended, Available, Not Found, and All).
EclWatch shows query status on a given cluster when a user
selects a cluster from the drop down list.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>