gh-1930 sendEmail rejecting some valid server addresses
The code for validating email addresses is incorrectly assuming that
the rules for a valid server address and the rules for a valid email domain
are the same. There is no good reason to validate the email server address
beyond passing it to the socket layer to resolve.

The rules for validating the domain part of the email address are also a little
over-strict and/or out of date - I think we really should remove much of that
validation and leave it to the smtp server to do, but that can be left for a
future fix. For now, just add support for - in email domain names.

Fixes gh-1930.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>