Allow dynamic file resolution in Roxie
Add dynamic flag to query factory, which activity
factories will use to define if dynamic file
resolution is to be turned on by default or just
specific to certain activities (TDXdynamicfilename).

Commoning up the variableFileName flag on all activities
to make sure we set the correct one and it gets
used on all activities that need it.

Implementing runOnce() on WorkUnit worker, so to
concentrate dynamic resolution to workunit execution
only, working for both dispatched and standalone
query types.

Seting the standalone/workunit payloads to have dynamic
file resolution, but still keeps deployed
queries as Roxie's standard resolve-on-startup.

Adding dynamicFiles flag to the query's property tree
when creating a standalone query package manager,
so the globalPackageSetManager propagates it to
QueryFactories when created.

Creating workunit when attaching without a WUID (which
happens with standalone queries), so we can use the
same worker for dynamic queries on both standalone
and one-off workunit queries.

Querying DLLs from a Workunit will also work if
it's a stand alone executable (the property tree holds
them in separate, but similar, places).