Implement URI backbone
The URI library deals with resolving URIs into files in
the file system. The current version contains a library
to parse and split the URI into recognizable parts, to
be used by the URI resolvers, that will come later.

This implementation uses the uriparser library (from
http://uriparser.sourceforge.net/ that is widely
available on Linux distros and MacOS back-ports, but
might have to be compiled on Windows). Further analysis
must be done to see if we include the source directly
or just distribute libraries on Windows (or find another
way).

There are some tests, but will be included separately,
since they require a refactoring in the old (dormant)
unittests tool.