Mark left dataset of join sorted even if global
Previously the function to indicate that the left hand side of a join
was sorted was never set for a global join.  However thor does correctly
support it if the dataset is globally sorted.
Also add flags to indicate if left/right are locally sorted to
potentially avoid some work in the sorting.

Signed-off-by: Gavin Halliday <gavin.halliday@lexisnexis.com>