gh-1832 Fix Deleting a workunit should remove from queues
This addresses an external users issue.  This shouldn't happen unless
the workunit state is no longer active, but it is somehow stuck in
the queue.

To test, one can simulate the issue by stopping the cluster, queueing
a workunit, setting the state to failed, then deleting the workunit.

Fixes gh-1832.
Signed-off-by: Anthony Fishbeck <Anthony.Fishbeck@lexisnexis.com>