Add password expiration reminder to ESP
When LDAP security is set and a user launches an ESP UI,
ESP checks how many days the user's password will expired.
If the expiration days is less than a pre-defined limit (
this limit is defined by configmgr for all of users using
the ESP), ESP will popup a message about the expiration.
If the user chooses to change the password, a password
update window will be displayed. If the user chooses not
to change, ESP may repeat the whole process in the next
day.

This fix needs the fix in https://github.com/hpcc-systems/
HPCC-Platform/pull/2035.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>