filesbasedn not being set in esp.xml
This fix was added to EE. Now, it should be CE also.
Existing code for generating the esp.xml sets the value of
filesbasedn based on dali setting. The filesbasedn was not set
in the esp.xml because it was not set in dali setting (dali
does not use filesbasedn, but ESP uses it). ESP should be able
to set the filesbasedn without dali setting. This fix in
esp_service_WsSMC.xsl sets the default value (the same value
as that in dali.xsd) if filesbasedn is not set. The fix in
environment.xml.in allows a user to set filesbasedn using
environment.xml.in. The fix in espsmcservice.xsd.in allows a
user to set filesbasedn using configmgr.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>