Thor could core in CGraphTempHandler::registerFile
Access to CGraphTempHandler is not threadsafe, and may core if there
are enough temp files registered in a job to cause the hash table to
be expanded (it's possible, though a lot less likely, to see issues even
with fewer temp files).

Add a CritSec to protect access to the hash table.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>