Fix gh-#2071 DFS exception when View Data File
When ldap is turned on and a user trys to view a data file,
the DFS exception '3 Lookup access denied for scope' shows.
The exception is thrown by the ViewFileWeb object when it
does a file look-up. Since the data file has been resolved
by WsDFU service, WsDFU service should pass the solved file
point to ViewFileWeb object and the ViewFileWeb may use the
point without doing the file loop-up. This fix adds a
method to the ViewFileWeb class. The method allows the
solved file point being given to the ViewFileWeb.

I also revise the code in WsDFU service for better
functionalities.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>