Roxie file cache was not checking timestamps
The code for checking and comparing filestamps when deciding if Roxie's
cached copy of a file was up to date seems to have been lost in the
big refactoring of package files, meaning that file mismatches would only
be detected when the size changed.

This had a few unfortunate effects:

1. Index changes are likely to be only detected by slaves but not by server
   nodes. This leads to 'partially suspended' queries which were are not
   giving helpful error messages.
2. If data is refreshed on thor, Roxie might continue to read a previously
   cached version of the data, even if Roxie was restarted.

See gh-2096. This is not a complete fix for all the issues there, since Roxie
is still falsely reporting that two versions of a data file are required
at the same time when all queries using the previous version have been
unloaded.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>